+++ backend/canisters/user_index/impl/src/lifecycle/init.rs
@@ -22,6 +22,7 @@ fn init(args: Args) {
         args.notifications_index_canister_id,
         args.identity_canister_id,
         args.proposals_bot_canister_id,
+        args.airdrop_bot_canister_id,
         args.cycles_dispenser_canister_id,
         args.storage_index_canister_id,
         args.escrow_canister_id,