+++ backend/integration_tests/src/setup.rs
@@ -134,6 +134,7 @@ fn install_canisters(env: &mut PocketIc, controller: Principal) -> CanisterIds {
         notifications_index_canister_id,
         identity_canister_id,
         proposals_bot_canister_id,
+        airdrop_bot_canister_id,
         cycles_dispenser_canister_id,
         storage_index_canister_id,
         escrow_canister_id,