+++ backend/tools/canister_installer/src/lib.rs
@@ -75,6 +75,7 @@ async fn install_service_canisters_impl(
         notifications_index_canister_id: canister_ids.notifications_index,
         identity_canister_id: canister_ids.identity,
         proposals_bot_canister_id: canister_ids.proposals_bot,
+        airdrop_bot_canister_id: canister_ids.airdrop_bot,
         storage_index_canister_id: canister_ids.storage_index,
         cycles_dispenser_canister_id: canister_ids.cycles_dispenser,
         escrow_canister_id: canister_ids.escrow,